+++ UI/src/gui/result/tabs/graph/GraphComponent.fxml
