+++ UI/src/gui/result/tabs/statistics/StatisticsComponent.fxml
